backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.1
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==10.4.0
PyJWT==2.10.1
pymongo==4.13.0
python-dotenv==1.1.0
setuptools==80.9.0
Werkzeug==3.1.3
wheel==0.45.1
